+++ lib/karafka/contracts/topic.rb
@@ -19,7 +19,6 @@ module Karafka
       required(:max_messages) { |val| val.is_a?(Integer) && val >= 1 }
       required(:initial_offset) { |val| %w[earliest latest].include?(val) }
       required(:max_wait_time) { |val| val.is_a?(Integer) && val >= 10 }
-      required(:manual_offset_management) { |val| [true, false].include?(val) }
       required(:name) { |val| val.is_a?(String) && Contracts::TOPIC_REGEXP.match?(val) }
       required(:subscription_group) { |val| val.nil? || (val.is_a?(String) && !val.empty?) }