+++ spec/lib/karafka/pro/base_consumer_spec.rb
@@ -1,9 +1,5 @@
 # frozen_string_literal: true
 
-require 'karafka/pro/base_consumer'
-require 'karafka/pro/routing/topic_extensions'
-require 'karafka/pro/processing/coordinator'
-
 RSpec.describe_current do
   subject(:consumer) do
     instance = working_class.new
@@ -20,7 +16,10 @@ RSpec.describe_current do
   let(:offset) { 123 }
   let(:topic) do
     build(:routing_topic).tap do |built|
-      built.singleton_class.prepend Karafka::Pro::Routing::TopicExtensions
+      [
+        Karafka::Pro::Routing::Features::VirtualPartitions::Topic,
+        Karafka::Pro::Routing::Features::LongRunningJob::Topic
+      ].each { |feature| built.singleton_class.prepend(feature) }
     end
   end
 
@@ -102,7 +101,7 @@ RSpec.describe_current do
     end
 
     context 'when everything went ok on consume with manual offset management' do
-      before { topic.manual_offset_management = true }
+      before { topic.manual_offset_management true }
 
       it { expect { consume_with_after.call }.not_to raise_error }
 
@@ -160,7 +159,7 @@ RSpec.describe_current do
 
     context 'when everything went ok on consume with automatic offset management' do
       before do
-        topic.manual_offset_management = false
+        topic.manual_offset_management false
         allow(client).to receive(:mark_as_consumed)
       end
 
@@ -181,7 +180,7 @@ RSpec.describe_current do
     end
 
     context 'when there was an error on consume with automatic offset management' do
-      before { topic.manual_offset_management = false }
+      before { topic.manual_offset_management false }
 
       let(:working_class) do
         ClassBuilder.inherit(described_class) do
@@ -222,7 +221,7 @@ RSpec.describe_current do
 
   describe '#on_before_enqueue for LRU' do
     before do
-      topic.long_running_job = true
+      topic.long_running_job true
       allow(client).to receive(:pause)
       consumer.messages = messages
     end
@@ -242,7 +241,7 @@ RSpec.describe_current do
     end
 
     before do
-      topic.long_running_job = true
+      topic.long_running_job true
       consumer.coordinator = coordinator
       consumer.client = client
       consumer.messages = messages
@@ -252,7 +251,7 @@ RSpec.describe_current do
     end
 
     context 'when everything went ok on consume with manual offset management' do
-      before { topic.manual_offset_management = true }
+      before { topic.manual_offset_management true }
 
       it { expect { consume_with_after.call }.not_to raise_error }
 
@@ -320,7 +319,7 @@ RSpec.describe_current do
 
     context 'when everything went ok on consume with automatic offset management' do
       before do
-        topic.manual_offset_management = false
+        topic.manual_offset_management false
         allow(client).to receive(:mark_as_consumed)
       end
 
@@ -341,7 +340,7 @@ RSpec.describe_current do
     end
 
     context 'when there was an error on consume with automatic offset management' do
-      before { topic.manual_offset_management = false }
+      before { topic.manual_offset_management false }
 
       let(:working_class) do
         ClassBuilder.inherit(described_class) do