+++ spec/lib/karafka/pro/contracts/base_spec.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-require 'karafka/pro/contracts/base'
-
-RSpec.describe_current do
-  it { expect(described_class).to be < ::Karafka::Contracts::Base }
-end