+++ spec/lib/karafka/pro/performance_tracker_spec.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require 'karafka/pro/performance_tracker'
-
 RSpec.describe_current do
   let(:tracker) { Karafka::Pro::PerformanceTracker.instance }